+++ pxd.c
@@ -1035,80 +1035,7 @@ static int pxd_init_disk(struct pxd_device *pxd_dev, unsigned int *blk_mq_queue_
 	  }
 
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(5,14,0)
-#ifdef __ELREPO9__
-#if LINUX_VERSION_CODE < KERNEL_VERSION(6,9,0)
-	  disk = blk_mq_alloc_disk(&pxd_dev->tag_set, pxd_dev);
-#else
-	  // WriteZero: enable only if capability is set AND fastpath is NOT enabled
-	  // (fastpath uses LVM which doesn't support discard/write_zeroes)
-	  unsigned int wz_sectors = (pxd_has_cap(pxd_dev, PXD_DEV_CAP_WRITE_ZEROES) && !fastpath_enabled(pxd_dev)) ?
-	                            (pxd_dev->discard_size / SECTOR_SIZE) : 0;
-	  struct queue_limits lim = {
-		  .logical_block_size = PXD_LBS,
-		  .physical_block_size = PXD_LBS,
-		  .max_segment_size = SEGMENT_SIZE,
-		  .max_segments = PXD_MAX_IO / PXD_LBS,
-		  .max_hw_sectors = PXD_MAX_IO / SECTOR_SIZE,
-		  .discard_alignment = pxd_dev->discard_granularity,
-		  .discard_granularity = pxd_dev->discard_granularity,
-		  .io_min = PXD_LBS,
-		  .io_opt = PXD_LBS,
-		  .max_hw_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
-		  .max_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
-		  // Control write_zeroes: enable if PXD_DEV_CAP_WRITE_ZEROES capability is set
-		  .max_write_zeroes_sectors = wz_sectors
-	  };
-	  disk = blk_mq_alloc_disk(&pxd_dev->tag_set, &lim, pxd_dev);
-#endif
-#elif defined(RHEL_RELEASE_CODE)
-#if RHEL_RELEASE_CODE >= RHEL_RELEASE_VERSION(9, 6)
-	  // WriteZero: enable only if capability is set AND fastpath is NOT enabled
-	  // (fastpath uses LVM which doesn't support discard/write_zeroes)
-	  unsigned int wz_sectors = (pxd_has_cap(pxd_dev, PXD_DEV_CAP_WRITE_ZEROES) && !fastpath_enabled(pxd_dev)) ?
-	                            (pxd_dev->discard_size / SECTOR_SIZE) : 0;
-	  struct queue_limits lim = {
-		  .logical_block_size = PXD_LBS,
-		  .physical_block_size = PXD_LBS,
-		  .max_segment_size = SEGMENT_SIZE,
-		  .max_segments = PXD_MAX_IO / PXD_LBS,
-		  .max_hw_sectors = PXD_MAX_IO / SECTOR_SIZE,
-		  .discard_alignment = pxd_dev->discard_granularity,
-		  .discard_granularity = pxd_dev->discard_granularity,
-		  .io_min = PXD_LBS,
-		  .io_opt = PXD_LBS,
-		  .max_hw_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
-		  .max_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
-		  // Control write_zeroes: enable if PXD_DEV_CAP_WRITE_ZEROES capability is set
-		  .max_write_zeroes_sectors = wz_sectors
-	  };
-	  disk = blk_mq_alloc_disk(&pxd_dev->tag_set, &lim, pxd_dev);
-#else
-	  disk = blk_mq_alloc_disk(&pxd_dev->tag_set, pxd_dev);
-#endif
-#elif LINUX_VERSION_CODE >= KERNEL_VERSION(6,9,0) || ((LINUX_VERSION_CODE >= KERNEL_VERSION(5,14,0) && defined(__EL8__)) && !defined(__ORACLE_UEK__))
-	  // WriteZero: enable only if capability is set AND fastpath is NOT enabled
-	  // (fastpath uses LVM which doesn't support discard/write_zeroes)
-	  unsigned int wz_sectors = (pxd_has_cap(pxd_dev, PXD_DEV_CAP_WRITE_ZEROES) && !fastpath_enabled(pxd_dev)) ?
-	                            (pxd_dev->discard_size / SECTOR_SIZE) : 0;
-	  struct queue_limits lim = {
-		  .logical_block_size = PXD_LBS,
-		  .physical_block_size = PXD_LBS,
-		  .max_segment_size = SEGMENT_SIZE,
-		  .max_segments = PXD_MAX_IO / PXD_LBS,
-		  .max_hw_sectors = PXD_MAX_IO / SECTOR_SIZE,
-		  .discard_alignment = pxd_dev->discard_granularity,
-		  .discard_granularity = pxd_dev->discard_granularity,
-		  .io_min = PXD_LBS,
-		  .io_opt = PXD_LBS,
-		  .max_hw_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
-		  .max_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
-		  // Control write_zeroes: enable if PXD_DEV_CAP_WRITE_ZEROES capability is set
-		  .max_write_zeroes_sectors = wz_sectors
-	  };
-	  disk = blk_mq_alloc_disk(&pxd_dev->tag_set, &lim, pxd_dev);
-#else
-	  disk = blk_mq_alloc_disk(&pxd_dev->tag_set, pxd_dev);
-#endif
+	  disk = pxd_alloc_disk(pxd_dev);
 
 	  if (IS_ERR(disk)) {
 		blk_mq_free_tag_set(&pxd_dev->tag_set);
