+++ pxd_compat.h
@@ -276,4 +276,82 @@ static inline void pxd_ida_remove(struct ida *ida, unsigned int id)
 	#endif
 }
 
+#ifdef __PX_BLKMQ__
+#include <linux/blk-mq.h>
+#include "pxd_core.h"
+
+static inline struct gendisk *pxd_alloc_disk(struct pxd_device *pxd_dev)
+{
+	struct gendisk *disk;
+
+#ifdef __ELREPO9__
+#if LINUX_VERSION_CODE < KERNEL_VERSION(6,9,0)
+	disk = blk_mq_alloc_disk(&pxd_dev->tag_set, pxd_dev);
+#else
+	unsigned int wz_sectors = (pxd_has_cap(pxd_dev, PXD_DEV_CAP_WRITE_ZEROES) && !fastpath_enabled(pxd_dev)) ?
+	                          (pxd_dev->discard_size / SECTOR_SIZE) : 0;
+	struct queue_limits lim = {
+		.logical_block_size = PXD_LBS,
+		.physical_block_size = PXD_LBS,
+		.max_segment_size = SEGMENT_SIZE,
+		.max_segments = PXD_MAX_IO / PXD_LBS,
+		.max_hw_sectors = PXD_MAX_IO / SECTOR_SIZE,
+		.discard_alignment = pxd_dev->discard_granularity,
+		.discard_granularity = pxd_dev->discard_granularity,
+		.io_min = PXD_LBS,
+		.io_opt = PXD_LBS,
+		.max_hw_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
+		.max_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
+		.max_write_zeroes_sectors = wz_sectors
+	};
+	disk = blk_mq_alloc_disk(&pxd_dev->tag_set, &lim, pxd_dev);
+#endif
+#elif defined(RHEL_RELEASE_CODE)
+#if RHEL_RELEASE_CODE >= RHEL_RELEASE_VERSION(9, 6)
+	unsigned int wz_sectors = (pxd_has_cap(pxd_dev, PXD_DEV_CAP_WRITE_ZEROES) && !fastpath_enabled(pxd_dev)) ?
+	                          (pxd_dev->discard_size / SECTOR_SIZE) : 0;
+	struct queue_limits lim = {
+		.logical_block_size = PXD_LBS,
+		.physical_block_size = PXD_LBS,
+		.max_segment_size = SEGMENT_SIZE,
+		.max_segments = PXD_MAX_IO / PXD_LBS,
+		.max_hw_sectors = PXD_MAX_IO / SECTOR_SIZE,
+		.discard_alignment = pxd_dev->discard_granularity,
+		.discard_granularity = pxd_dev->discard_granularity,
+		.io_min = PXD_LBS,
+		.io_opt = PXD_LBS,
+		.max_hw_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
+		.max_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
+		.max_write_zeroes_sectors = wz_sectors
+	};
+	disk = blk_mq_alloc_disk(&pxd_dev->tag_set, &lim, pxd_dev);
+#else
+	disk = blk_mq_alloc_disk(&pxd_dev->tag_set, pxd_dev);
+#endif
+#elif LINUX_VERSION_CODE >= KERNEL_VERSION(6,9,0) || ((LINUX_VERSION_CODE >= KERNEL_VERSION(5,14,0) && defined(__EL8__)) && !defined(__ORACLE_UEK__))
+	unsigned int wz_sectors = (pxd_has_cap(pxd_dev, PXD_DEV_CAP_WRITE_ZEROES) && !fastpath_enabled(pxd_dev)) ?
+	                          (pxd_dev->discard_size / SECTOR_SIZE) : 0;
+	struct queue_limits lim = {
+		.logical_block_size = PXD_LBS,
+		.physical_block_size = PXD_LBS,
+		.max_segment_size = SEGMENT_SIZE,
+		.max_segments = PXD_MAX_IO / PXD_LBS,
+		.max_hw_sectors = PXD_MAX_IO / SECTOR_SIZE,
+		.discard_alignment = pxd_dev->discard_granularity,
+		.discard_granularity = pxd_dev->discard_granularity,
+		.io_min = PXD_LBS,
+		.io_opt = PXD_LBS,
+		.max_hw_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
+		.max_discard_sectors = pxd_dev->discard_size / SECTOR_SIZE,
+		.max_write_zeroes_sectors = wz_sectors
+	};
+	disk = blk_mq_alloc_disk(&pxd_dev->tag_set, &lim, pxd_dev);
+#else
+	disk = blk_mq_alloc_disk(&pxd_dev->tag_set, pxd_dev);
+#endif
+
+	return disk;
+}
+#endif /* __PX_BLKMQ__ */
+
 #endif //GDFS_PXD_COMPAT_H
